test AAK over the parameter space
